refactor: split telegram delivery and unify media/frontmatter/i18n pipelines
